+++ IteaIOSDip/Authorization/AuthenticationViewController+Extension.swift
@@ -9,6 +9,94 @@
 import Foundation
 import UIKit
 
+// MARK: - AUTHENTICATION - MAIL CHECK -
+extension AuthenticationViewController{
+    
+    func studentAuthentication(students : [IteaStudent]) -> IteaStudent {
+        var loginStudent = IteaStudent()
+        for user in students {
+            if user.mail == userNameTextField.text {
+                if user.password == userPasswordTextField.text {
+                    loginStudent = user
+                }
+            }
+        }
+        return loginStudent
+    }
+}
+
+//  MARK: - VALIDATION -
+extension AuthenticationViewController {
+    
+    func showValidationErrors(error: ValidationErrors.ErrorTextEnum, errorLabel: UILabel, subLabel: NSLayoutConstraint) {
+        subLabel.priority = UILayoutPriority(rawValue: 900)
+        errorLabel.text = validationErrors.errorKey(error)
+    }
+    
+    func validateLogin(inputLogin: String, inputPassword: String) -> Bool {
+        var validateFields = false
+        if validation.validateMail(mail: inputLogin) == false {
+            emailValidationErrorTextLabel.text = validationErrors.errorKey(.invalidEmail)
+        }
+        if validation.validatePassword(password: inputPassword) == false {
+            passwordValidationErrorTextLabel.text = validationErrors.errorKey(.invalidPassword)
+        }
+        if validation.validateMail(mail: inputLogin) == true && validation.validatePassword(password: inputPassword) == true {
+            validateFields = true
+        }
+        return validateFields
+    }
+}
+
+//  MARK: - TEXT FIELDS DELEGATE -
+extension AuthenticationViewController: UITextFieldDelegate {
+    func textFieldShouldBeginEditing(_ textField: UITextField) -> Bool {
+        bottomHeightConstraint.constant = 250
+        return true
+    }
+    
+    func textFieldShouldReturn(_ textField: UITextField) -> Bool {
+        switch textField {
+        case userNameTextField:
+            userPasswordTextField.becomeFirstResponder()
+        case userPasswordTextField:
+            userPasswordTextField.resignFirstResponder()
+            bottomHeightConstraint.constant = 0
+        default:
+            userNameTextField.becomeFirstResponder()
+        }
+        return true
+    }
+       
+    func textFieldDidBeginEditing(_ textField: UITextField) {
+        switch textField {
+        case userNameTextField:
+            showLineRespontMailFieldHeightConstraint.priority = UILayoutPriority(rawValue: 900)
+            hideValidationErrors()
+        case userPasswordTextField:
+            userPasswordTextField.text = ""
+            showMassonsEyeImageView.image = UIImage(systemName: "eye.slash")
+            showLineRespondPasswordFieldHeightConstraint.priority = UILayoutPriority(rawValue: 900)
+            hideValidationErrors()
+        default:
+            hideValidationErrors()
+        }
+    }
+    
+    func textFieldDidEndEditing(_ textField: UITextField) {
+        switch textField {
+        case userNameTextField:
+            showLineRespontMailFieldHeightConstraint.priority = UILayoutPriority(rawValue: 600)
+        case userPasswordTextField:
+            showLineRespondPasswordFieldHeightConstraint.priority = UILayoutPriority(rawValue: 600)
+        default:
+            showLineRespondPasswordFieldHeightConstraint.priority = UILayoutPriority(rawValue: 600)
+            showLineRespontMailFieldHeightConstraint.priority = UILayoutPriority(rawValue: 600)
+        }
+    }
+}
+
+// MARK: - FIELDS VALIDATION -
 extension AuthenticationViewController {
     
     func hideValidationErrors() {
@@ -32,64 +120,56 @@ extension AuthenticationViewController {
     }
 }
 
+//  MARK: - USER INTERFACE UI -
 extension AuthenticationViewController {
-    
-        func updateAuthUi() {
-            
-            iteaLogoMainView.layer.borderColor = UIColor.white.cgColor
-            iteaLogoMainView.layer.borderWidth = 2.0
-            iteaLogoIView.layer.borderColor = UIColor.white.cgColor
-            iteaLogoIView.layer.borderWidth = 2.0
-            iteaLogoTView.layer.borderColor = UIColor.white.cgColor
-            iteaLogoTView.layer.borderWidth = 2.0
-            iteaLogoEView.layer.borderColor = UIColor.white.cgColor
-            iteaLogoEView.layer.borderWidth = 2.0
-            
-
-            
-            let blurEffect = UIBlurEffect(style: .prominent)
-                let logInBlurView = UIVisualEffectView(effect: blurEffect)
-            let sighnUpBlurView = UIVisualEffectView(effect: blurEffect)
-                logInBlurView.translatesAutoresizingMaskIntoConstraints = false
-            sighnUpBlurView.translatesAutoresizingMaskIntoConstraints = false
-            logInButtonView.insertSubview(logInBlurView, at: 0)
-            signUpButtonView.insertSubview(sighnUpBlurView, at: 0)
-            logInBlurView.clipsToBounds = true
-            sighnUpBlurView.clipsToBounds = true
-            
-                logInBlurView.layer.cornerRadius = 12.0
-                sighnUpBlurView.layer.cornerRadius = 12.0
-                NSLayoutConstraint.activate([sighnUpBlurView.heightAnchor.constraint(equalTo: signUpButtonView.heightAnchor),
-                                             sighnUpBlurView.widthAnchor.constraint(equalTo: signUpButtonView.widthAnchor), logInBlurView.heightAnchor.constraint(equalTo: logInButtonView.heightAnchor), logInBlurView.widthAnchor.constraint(equalTo: logInButtonView.widthAnchor)
-                ])
-            
-            signUpButtonView.clipsToBounds = true
-            signUpButtonView.layer.masksToBounds = false
-            signUpButtonView.layer.shadowColor = UIColor.black.cgColor
-            signUpButtonView.layer.shadowOpacity = 0.5
-            signUpButtonView.layer.shadowOffset = CGSize(width: 10, height: 10)
-            signUpButtonView.layer.shadowRadius = 12.0
-            
-            logInButtonView.clipsToBounds = true
-            logInButtonView.layer.masksToBounds = false
-            logInButtonView.layer.shadowColor = UIColor.black.cgColor
-            logInButtonView.layer.shadowOpacity = 0.5
-            logInButtonView.layer.shadowOffset = CGSize(width: 10, height: 10)
-            logInButtonView.layer.shadowRadius = 12.0
-            
-            
-//            let backgroundImage = UIImageView(frame: UIScreen.main.bounds)
-//            backgroundImage.image = UIImage(named: "backgroundlogin")
-//            backgroundImage.contentMode = .scaleAspectFill
-//            let backgroundForImage = UIView(frame: UIScreen.main.bounds)
-//            backgroundForImage.backgroundColor = UIColor.red
-//            backgroundForImage.alpha = 0.8
-//            view.addSubview(backgroundImage)
-//            view.addSubview(backgroundForImage)
-//            self.view.sendSubviewToBack(backgroundForImage)
-//            self.view.sendSubviewToBack(backgroundImage)
-            
-            
+    func updateAuthUi() {
+        iteaLogoMainView.layer.borderColor = UIColor.white.cgColor
+        iteaLogoMainView.layer.borderWidth = 2.0
+        
+        iteaLogoIView.layer.borderColor = UIColor.white.cgColor
+        iteaLogoIView.layer.borderWidth = 2.0
+        iteaLogoTView.layer.borderColor = UIColor.white.cgColor
+        iteaLogoTView.layer.borderWidth = 2.0
+        iteaLogoEView.layer.borderColor = UIColor.white.cgColor
+        
+        iteaLogoEView.layer.borderWidth = 2.0
+        signUpButtonView.clipsToBounds = true
+        signUpButtonView.layer.masksToBounds = false
+        signUpButtonView.layer.shadowColor = UIColor.black.cgColor
+        signUpButtonView.layer.shadowOpacity = 0.5
+        signUpButtonView.layer.shadowOffset = CGSize(width: 10, height: 10)
+        signUpButtonView.layer.shadowRadius = 12.0
+        
+        logInButtonView.clipsToBounds = true
+        logInButtonView.layer.masksToBounds = false
+        logInButtonView.layer.shadowColor = UIColor.black.cgColor
+        logInButtonView.layer.shadowOpacity = 0.5
+        logInButtonView.layer.shadowOffset = CGSize(width: 10, height: 10)
+        logInButtonView.layer.shadowRadius = 12.0
     }
+}
 
+//  MARK: - SHOW HIDE PASSWORD BUTTON -
+extension AuthenticationViewController {
+    func showPassword() {
+        if userPasswordTextField.isSecureTextEntry == true {
+            userPasswordTextField.isSecureTextEntry = false
+            showHidePasswordButton.setTitle("HIDE", for: .normal)
+            showMassonsEyeImageView.image = UIImage(systemName: "eye.slash")
+        } else {
+            userPasswordTextField.isSecureTextEntry = true
+            showHidePasswordButton.setTitle("SHOW", for: .normal)
+            showMassonsEyeImageView.image = UIImage(systemName: "eye")
+        }
+    }
+}
+
+//  MARK: - ALERT ERRORS ACTION
+extension AuthenticationViewController {
+    func showAlertErrorAction(title: String, message: String) {
+        let alertController = UIAlertController(title: title, message: message, preferredStyle: .alert)
+        let alertAction = UIAlertAction(title: "OK", style: .cancel) { (_) in}
+        alertController.addAction(alertAction)
+        present(alertController, animated: true, completion: nil)
+    }
 }
