+++ IteaIOSDip/Filter/FilterViewController.swift
@@ -48,13 +48,12 @@ class FilterViewController: UIViewController {
     
     override func viewDidLoad() {
         super.viewDidLoad()
-
 //      recieve course data
         setDefaultsFilterCourse(courses: flow)
-
 //      set defaults buttons tap FUNC
         defaultFilterButtons()
         uidesignFilter()
+        saveFilterButton.iteaButton()
 
 //  MARK: - MAKE RADIO BUTTONS
         let radioButtons: [UIView] = [nightCheckExternalView,
@@ -140,156 +139,5 @@ class FilterViewController: UIViewController {
             filterAlert(title: "Ooops!", message: "Не найдено курсов с такими параметрами")
             defaultFilterButtons()
         }
-    }
-}
-
-extension FilterViewController {
-
-    
-//    TODO: - MAKE FUNCTIOM FOR BUTTONS CHANGE
-//    func chngeRadioButtons(activeFilter: Bool, notActive: Bool) {
-//        var active: CGColor = UIColor.black.cgColor
-//        var notActive: CGColor = UIColor.white.cgColor
-//        
-//    
-//        if
-//        
-//        allFlowCheckInternalView.layer.backgroundColor = UIColor.black.cgColor
-//        programmingCheckInternalView.layer.backgroundColor = UIColor.clear.cgColor
-//        frontEndCheckInternalView.layer.backgroundColor = UIColor.clear.cgColor
-//        designCheckInternalView.layer.backgroundColor = UIColor.clear.cgColor
-//        qaCheckInternalView.layer.backgroundColor = UIColor.clear.cgColor
-//        otherCheckInternalView.layer.backgroundColor = UIColor.clear.cgColor
-//    }
-}
-
-
-extension FilterViewController {
-    
-    @objc func didNightTapAction() {
-        filterSaved(state: false)
-        sortbyDate = []
-        nightCheckInternalView.layer.backgroundColor = UIColor.white.cgColor
-        dayCheckInternalView.layer.backgroundColor = UIColor.clear.cgColor
-
-        for  flowItems in flow {
-            if flowItems.courseTime == "night" {
-                sortbyDate.append(flowItems)
-            }
-        }
-    }
-    
-    @objc func didDayTapAction() {
-        filterSaved(state: false)
-        sortbyDate = []
-        dayCheckInternalView.layer.backgroundColor = UIColor.white.cgColor
-        nightCheckInternalView.layer.backgroundColor = UIColor.clear.cgColor
-        
-        for flowItems in flow {
-            if flowItems.courseTime == "day" {
-                sortbyDate.append(flowItems)
-            }
-        }
-    }
-    
-    @objc func didallFlowAction() {
-        filterSaved(state: false)
-        sortbyType = []
-        allFlowCheckInternalView.layer.backgroundColor = UIColor.white.cgColor
-        programmingCheckInternalView.layer.backgroundColor = UIColor.clear.cgColor
-        frontEndCheckInternalView.layer.backgroundColor = UIColor.clear.cgColor
-        designCheckInternalView.layer.backgroundColor = UIColor.clear.cgColor
-        qaCheckInternalView.layer.backgroundColor = UIColor.clear.cgColor
-        otherCheckInternalView.layer.backgroundColor = UIColor.clear.cgColor
-        
-        for flowItems in flow {
-                sortbyType.append(flowItems)
-        }
-    }
-    
-    @objc func didProgramingAction() {
-        filterSaved(state: false)
-        sortbyType = []
-        allFlowCheckInternalView.layer.backgroundColor = UIColor.clear.cgColor
-        programmingCheckInternalView.layer.backgroundColor = UIColor.white.cgColor
-        frontEndCheckInternalView.layer.backgroundColor = UIColor.clear.cgColor
-        designCheckInternalView.layer.backgroundColor = UIColor.clear.cgColor
-        qaCheckInternalView.layer.backgroundColor = UIColor.clear.cgColor
-        otherCheckInternalView.layer.backgroundColor = UIColor.clear.cgColor
-        
-        for flowItems in flow {
-            if flowItems.courseType == "programming" {
-                sortbyType.append(flowItems)
-            }
-        }
-    }
-    
-    @objc func didFrontEndAction() {
-        filterSaved(state: false)
-        sortbyType = []
-        allFlowCheckInternalView.layer.backgroundColor = UIColor.clear.cgColor
-        programmingCheckInternalView.layer.backgroundColor = UIColor.clear.cgColor
-        frontEndCheckInternalView.layer.backgroundColor = UIColor.white.cgColor
-        designCheckInternalView.layer.backgroundColor = UIColor.clear.cgColor
-        qaCheckInternalView.layer.backgroundColor = UIColor.clear.cgColor
-        otherCheckInternalView.layer.backgroundColor = UIColor.clear.cgColor
-        
-        
-        for flowItems in flow {
-            if flowItems.courseType == "front" {
-                sortbyType.append(flowItems)
-            }
-        }
-    }
-    
-    @objc func didDesignAction() {
-        filterSaved(state: false)
-        sortbyType = []
-        allFlowCheckInternalView.layer.backgroundColor = UIColor.clear.cgColor
-        programmingCheckInternalView.layer.backgroundColor = UIColor.clear.cgColor
-        frontEndCheckInternalView.layer.backgroundColor = UIColor.clear.cgColor
-        designCheckInternalView.layer.backgroundColor = UIColor.white.cgColor
-        qaCheckInternalView.layer.backgroundColor = UIColor.clear.cgColor
-        otherCheckInternalView.layer.backgroundColor = UIColor.clear.cgColor
-        
-        for flowItems in flow {
-            if flowItems.courseType == "design" {
-                sortbyType.append(flowItems)
-            }
-        }
-    }
-    
-    @objc func didQaAction() {
-        filterSaved(state: false)
-        sortbyType = []
-        allFlowCheckInternalView.layer.backgroundColor = UIColor.clear.cgColor
-        programmingCheckInternalView.layer.backgroundColor = UIColor.clear.cgColor
-        frontEndCheckInternalView.layer.backgroundColor = UIColor.clear.cgColor
-        designCheckInternalView.layer.backgroundColor = UIColor.clear.cgColor
-        qaCheckInternalView.layer.backgroundColor = UIColor.white.cgColor
-        otherCheckInternalView.layer.backgroundColor = UIColor.clear.cgColor
-
-        for flowItems in flow {
-            if flowItems.courseType == "qa" {
-                sortbyType.append(flowItems)
-            }
-        }
-    }
-    
-    @objc func didOtherAction() {
-        filterSaved(state: false)
-        sortbyType = []
-        allFlowCheckInternalView.layer.backgroundColor = UIColor.clear.cgColor
-        programmingCheckInternalView.layer.backgroundColor = UIColor.clear.cgColor
-        frontEndCheckInternalView.layer.backgroundColor = UIColor.clear.cgColor
-        designCheckInternalView.layer.backgroundColor = UIColor.clear.cgColor
-        qaCheckInternalView.layer.backgroundColor = UIColor.clear.cgColor
-        otherCheckInternalView.layer.backgroundColor = UIColor.white.cgColor
-        
-        for flowItems in flow {
-            if flowItems.courseType == "other" {
-                sortbyType.append(flowItems)
-            }
-        }        
     }
 }
