+++ IteaIOSDip/Profile/ProfileViewController.swift
@@ -69,6 +69,7 @@ class ProfileViewController: UIViewController {
         updateProfileUI()
         lastCourseButton.iteaButton()
         logOutButton.iteaButton()
+        profileInfoView.blurbuttons()
         
         nameTextField.delegate = self
         lastNameTextField.delegate = self
@@ -123,6 +124,7 @@ class ProfileViewController: UIViewController {
             }
         }
     }
+    
     @IBAction func didTapCancelActionButton(_ sender: Any) {
         editProfile = false
         editImageView.isHidden = false
@@ -154,269 +156,11 @@ class ProfileViewController: UIViewController {
         logOutAlert.addAction(cancelAction)
         present(logOutAlert, animated: true, completion: nil)
     }
-
-    
-    func saveStudentInfo() {
-        profile.name = nameTextField.text
-        profile.lastName = lastNameTextField.text
-        profile.age = Int(ageTextField.text ?? "0")
-        profile.city = cityTextField.text
-        profile.birthday = birhdayTextField.text
-        profile.mail = emailTextField.text
-        profile.phoneNumber = phoneTextField.text
-        profile.studenCurrentCourse = courentCourseTextField.text
-        profile.workPlace = workPlaceTextField.text
-    }
-    
-    func updateStudentInfo(profile: IteaStudent, edit: Bool, style: UITextField.BorderStyle, color: UIColor) {
-        
-        nameTextField.text = profile.name
-        lastNameTextField.text = profile.lastName
-        ageTextField.text = "\(profile.age ?? 0)"
-        cityTextField.text = profile.city
-        birhdayTextField.text = profile.birthday
-        emailTextField.text = profile.mail
-        phoneTextField.text = profile.phoneNumber
-        courentCourseTextField.text = profile.studenCurrentCourse
-        workPlaceTextField.text = profile.workPlace
-        
-        nameTextField.isEnabled = edit
-        lastNameTextField.isEnabled = edit
-        lastNameTextField.textColor = UIColor.black
-        ageTextField.isEnabled = edit
-        cityTextField.isEnabled = edit
-        birhdayTextField.isEnabled = edit
-        emailTextField.isEnabled = edit
-        phoneTextField.isEnabled = edit
-        courentCourseTextField.isEnabled = edit
-        workPlaceTextField.isEnabled = edit
-        
-        nameTextField.borderStyle = style
-        nameTextField.backgroundColor = color
-        lastNameTextField.borderStyle = style
-        lastNameTextField.backgroundColor = color
-        ageTextField.borderStyle = style
-        ageTextField.backgroundColor = color
-        cityTextField.borderStyle = style
-        cityTextField.backgroundColor = color
-        birhdayTextField.borderStyle = style
-        birhdayTextField.backgroundColor = color
-        emailTextField.borderStyle = style
-        emailTextField.backgroundColor = color
-        phoneTextField.borderStyle = style
-        phoneTextField.backgroundColor = color
-        courentCourseTextField.borderStyle = style
-        courentCourseTextField.backgroundColor = color
-        workPlaceTextField.borderStyle = style
-        workPlaceTextField.backgroundColor = color
-        
-        if nameTextField.isEnabled == true {
-            nameTextField.textColor = UIColor.black
-        } else {
-            nameTextField.textColor = UIColor.white
-        }
-        if lastNameTextField.isEnabled == true {
-            lastNameTextField.textColor = UIColor.black
-        } else {
-            lastNameTextField.textColor = UIColor.white
-        }
-    }
 }
 
 extension ProfileViewController {
-    
     @objc func keyboardWillHide() {
         bottomHeightConstraint.constant = 0
         self.view.endEditing(true)
-    }
-}
-
-extension ProfileViewController: UITextFieldDelegate {
-    
-    func textFieldShouldBeginEditing(_ textField: UITextField) -> Bool {
-        bottomHeightConstraint.constant = 270
-        return true
-    }
-    
-
-    
-    func textFieldShouldReturn(_ textField: UITextField) -> Bool {
-        
-        switch textField {
-        case nameTextField:
-            lastNameTextField.becomeFirstResponder()
-        case lastNameTextField:
-            ageTextField.becomeFirstResponder()
-        case ageTextField:
-            cityTextField.becomeFirstResponder()
-        case cityTextField:
-            birhdayTextField.becomeFirstResponder()
-        case birhdayTextField:
-            emailTextField.becomeFirstResponder()
-        case emailTextField:
-            phoneTextField.becomeFirstResponder()
-        case phoneTextField:
-            courentCourseTextField.becomeFirstResponder()
-        case courentCourseTextField:
-            workPlaceTextField.becomeFirstResponder()
-        case workPlaceTextField:
-            workPlaceTextField.resignFirstResponder()
-             bottomHeightConstraint.constant = 0
-        default:
-            nameTextField.becomeFirstResponder()
-        }
-        return true
-    }
-    
-    func textFieldDidBeginEditing(_ textField: UITextField) {
-        switch textField {
-        case nameTextField:
-            nameValidationErrorHeightConstraint.priority = UILayoutPriority(rawValue: 600)
-            nameValidationTextLabel.text = ""
-        case lastNameTextField:
-            lastNameValidationHeightConstraint.priority = UILayoutPriority(rawValue: 600)
-            lastNameValidationTextLabel.text = ""
-        case emailTextField:
-            emailValidationErrorHeightConstraint.priority = UILayoutPriority(rawValue: 600)
-            emailValidationTextLabel.text = ""
-        case phoneTextField:
-            phoneValidationErrorHeightConstraint.priority = UILayoutPriority(rawValue: 600)
-            phoneValidationTextLabel.text = ""
-        default:
-            nameValidationErrorHeightConstraint.priority = UILayoutPriority(rawValue: 600)
-        }
-        
-    }
-    
-    
-}
-
-
-extension ProfileViewController {
-    
-    func validateEditProfile(inputName: String, inputLastName: String, inputMail: String, inputPhone: String) -> Bool {
-        var validateFields = false
-        
-        if validation.validateName(name: inputName) == false {
-            nameValidationTextLabel.text = validationErrors.errorKey(.invalidFirstName) + validationErrors.errorKey(.validateName)
-            nameValidationErrorHeightConstraint.priority = UILayoutPriority(rawValue: 900)
-        }
-        if validation.validateLastNAme(lastName: inputLastName) == false {
-            lastNameValidationTextLabel.text = validationErrors.errorKey(.invalidLastName) + validationErrors.errorKey(.validateLastName)
-            lastNameValidationHeightConstraint.priority = UILayoutPriority(rawValue: 900)
-        }
-        if validation.validateMail(mail: inputMail) == false {
-            emailValidationTextLabel.text = validationErrors.errorKey(.invalidEmail) + validationErrors.errorKey(.validateMail)
-            emailValidationErrorHeightConstraint.priority = UILayoutPriority(rawValue: 900)
-        }
-        if validation.validatePhone(phone: inputPhone) == false {
-            phoneValidationTextLabel.text = validationErrors.errorKey(.invalidPhoneNumber) + validationErrors.errorKey(.validatePhone)
-            phoneValidationErrorHeightConstraint.priority = UILayoutPriority(rawValue: 900)
-        }
-        if validation.validateName(name: inputName) == true && validation.validateLastNAme(lastName: inputLastName) == true &&  validation.validateMail(mail: inputMail) == true && validation.validatePhone(phone: inputPhone) == true {
-            validateFields = true
-        }
-        return validateFields
-        
-    }
-    
-    
-    
-    
-    
-    
-}
-
-
-extension ProfileViewController {
-    func showAlertErrorAction(title: String, message: String) {
-        let alertController = UIAlertController(title: title, message: message, preferredStyle: .alert)
-        let alertAction = UIAlertAction(title: "OK", style: .cancel) { (_) in}
-        alertController.addAction(alertAction)
-        present(alertController, animated: true, completion: nil)
-    }
-}
-
-
-
-extension ProfileViewController {
-    
-    func updateProfileUI() {
-        
-        
-        viewForStudentView.clipsToBounds = true
-        viewForStudentView.layer.masksToBounds = false
-        viewForStudentView.layer.cornerRadius = viewForStudentView.frame.width / 2
-        viewForStudentView.layer.shadowColor = UIColor.black.cgColor
-        viewForStudentView.layer.shadowOffset = CGSize(width: 3, height: 3)
-        viewForStudentView.layer.shadowRadius = 12.0
-        viewForStudentView.layer.shadowOpacity = 0.4
-        
-        studentUserBlurImageView.image = UIImage(named: profile.userPicture ?? "noCourseImage")
-        studentUserBlurImageView.contentMode = .scaleAspectFill
-        studentUserBlurImageView.alpha = 0.1
-        
-//        let blurEffect = UIBlurEffect(style: UIBlurEffect.Style.systemUltraThinMaterialLight)
-//        let blurEffectView = UIVisualEffectView(effect: blurEffect)
-//               blurEffectView.frame = view.bounds
-//               blurEffectView.autoresizingMask = [.flexibleWidth, .flexibleHeight]
-//
-//
-//               let backgroundImage = UIImageView(frame: UIScreen.main.bounds)
-//               backgroundImage.image = UIImage(named: backgroundPictureFromUser)
-//               backgroundImage.contentMode = .scaleAspectFill
-//
-//               view.addSubview(blurEffectView)
-//               view.addSubview(backgroundImage)
-//               self.view.sendSubviewToBack(blurEffectView)
-//               self.view.sendSubviewToBack(backgroundImage)
-
-//        lastCourseButton.clipsToBounds = true
-//        lastCourseButton.layer.backgroundColor = UIColor(red: 111/255, green: 169/255, blue: 145/255, alpha: 1).cgColor
-//        lastCourseButton.layer.cornerRadius = 12.0
-//        lastCourseButton.isOpaque = true
-//        lastCourseButton.layer.masksToBounds = false
-//        lastCourseButton.layer.shadowColor = UIColor.black.cgColor
-//        lastCourseButton.layer.shadowOpacity = 0.5
-//        lastCourseButton.layer.shadowOffset = CGSize(width: 10, height: 10)
-//        lastCourseButton.layer.shadowRadius = 12.0
-        
-        //
-
-        studentImageView.image = UIImage(named: profile.userPicture ?? "noCourseImage")
-        studentImageView.clipsToBounds = true
-        studentImageView.contentMode = .scaleAspectFill
-        studentImageView.layer.cornerRadius = studentImageView.frame.width / 2
-        studentImageView.layer.borderColor = UIColor.red.cgColor
-        studentImageView.layer.borderWidth = 2
-
-        nameStudentView.clipsToBounds = true
-        nameStudentView.layer.masksToBounds = false
-        nameStudentView.layer.cornerRadius = 12.0
-        nameStudentView.layer.shadowColor = UIColor.black.cgColor
-        nameStudentView.layer.shadowOffset = CGSize(width: 10, height: 10)
-        nameStudentView.layer.shadowRadius = 30.0
-        nameStudentView.layer.shadowOpacity = 0.4
-        
-//        profileInfoView
-        
-        let blurEffect = UIBlurEffect(style: .systemThickMaterialLight)
-        let blurView = UIVisualEffectView(effect: blurEffect)
-        blurView.translatesAutoresizingMaskIntoConstraints = false
-        blurView.clipsToBounds = true
-        blurView.layer.cornerRadius = 12
-        profileInfoView.insertSubview(blurView, at: 0)
-
-        NSLayoutConstraint.activate([blurView.heightAnchor.constraint(equalTo: profileInfoView.heightAnchor), blurView.widthAnchor.constraint(equalTo: profileInfoView.widthAnchor),
-        ])
-        
-        profileInfoView.layer.masksToBounds = false
-        profileInfoView.layer.shadowColor = UIColor.black.cgColor
-        profileInfoView.layer.shadowOpacity = 0.3
-        profileInfoView.layer.shadowOffset = CGSize(width: 10, height: 10)
-        profileInfoView.layer.shadowRadius = 6.0
-        
-        
-        
     }
 }
